pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cheragh"
version = "1.0.0"
description = "Composable RAG toolkit for standard Python applications"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RAG Toolkit Maintainers" }]
keywords = ["rag", "retrieval", "llm", "embeddings", "information-retrieval"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["numpy>=1.24", "pydantic>=2"]

[project.optional-dependencies]
local = ["sentence-transformers>=2.2"]
openai = ["openai>=1.0"]
cohere = ["cohere>=5.0"]
voyage = ["voyageai>=0.2"]
anthropic = ["anthropic>=0.25"]
litellm = ["litellm>=1.0"]
rerank = ["sentence-transformers>=2.2"]
raptor = ["scikit-learn>=1.0"]
bm25 = ["rank-bm25>=0.2.2"]
config = ["PyYAML>=6.0", "pydantic>=2"]
pdf = ["pypdf>=4.0"]
docx = ["python-docx>=1.0"]
faiss = ["faiss-cpu>=1.7"]
chroma = ["chromadb>=0.5"]
qdrant = ["qdrant-client>=1.9"]
fastapi = ["fastapi>=0.110", "uvicorn>=0.29", "pydantic>=2"]
redis = ["redis>=5.0"]
dev = ["pytest>=7", "pytest-cov", "ruff", "mypy", "build", "twine", "PyYAML>=6.0"]
all = [
  "sentence-transformers>=2.2",
  "openai>=1.0",
  "cohere>=5.0",
  "voyageai>=0.2",
  "anthropic>=0.25",
  "litellm>=1.0",
  "scikit-learn>=1.0",
  "rank-bm25>=0.2.2",
  "PyYAML>=6.0",
  "pydantic>=2",
  "pypdf>=4.0",
  "python-docx>=1.0",
  "faiss-cpu>=1.7",
  "chromadb>=0.5",
  "qdrant-client>=1.9",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "redis>=5.0",
]

[project.scripts]
cheragh = "cheragh.cli.main:cli_entrypoint"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
cheragh = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
